pep8 fixes

+++ jd_lib/datadb.py
@@ -1,5 +1,11 @@
 #!/usr/bin/env python3
 
+"""
+Simple interface to a minimalistic data store.
+
+Data is stored in a table in SQLite3.
+"""
+
 import sqlite3
 from pprint import pprint
 
@@ -9,6 +15,7 @@ DEBUGIT = False  # True # False
 
 
 def mydebug(*data):
+    """Print extra info when DEBG is enabled."""
     if DEBUGIT:
         pprint(data)
     return True
@@ -17,17 +24,18 @@ def mydebug(*data):
 
 
 def dict_factory(cursor, row) -> dict:
-    d = {}
+    """Map function. Make result rows accessible via column name."""
+    data = {}
     for idx, col in enumerate(cursor.description):
-        d[col[0]] = row[idx]
-    return d
+        data[col[0]] = row[idx]
+    return data
 
 
-class DataDB(object):
-    _db = None
-    _table = None
+class DataDB():
+    """Database class to contain all the necessary parameters and functions."""
 
     def __init__(self, uri=None, table=None):
+        """Update internal class default values."""
         self._db = uri or 'app.db'
         self._table = table or 'resource'
         try:
@@ -51,8 +59,8 @@ class DataDB(object):
                 );'''.format(self._table)
         cur.execute(query)
         query = 'SELECT id FROM {};'.format(self._table)
-        id = cur.execute(query).fetchone()
-        if not id:
+        row = cur.execute(query).fetchone()
+        if not row:
             query = 'INSERT INTO {} (id, name, value) '.format(self._table) + \
                       'VALUES (?,?,?)'
             cur.execute(query, [0, 'dummy name', 'dummy value'])
@@ -60,21 +68,26 @@ class DataDB(object):
         conn.close()
         return True
 
-    def data_get_byid(self, id=None) -> dict:
+    def data_get_byid(self, key_id=None) -> dict:
+        """Fetch a specific row or all rows from the database."""
         conn = sqlite3.connect(self._db, uri=True)
         conn.row_factory = dict_factory
         cur = conn.cursor()
-        if id is None:
+        if key_id is None:
             query = 'SELECT * FROM {};'.format(self._table)
             data = cur.execute(query).fetchall()
         else:
             query = 'SELECT * FROM {} WHERE id=?;'.format(self._table)
-            data = cur.execute(query, [id, ]).fetchone()
+            data = cur.execute(query, [key_id, ]).fetchone()
         conn.close()
         mydebug("data", data)
         return data
 
-    def data_get_byname(self, name) -> dict:
+    def data_get_byname(self, name=None) -> dict:
+        """Fetch a specific row or all rows from the database."""
+        if name is None:
+            mydebug("No name supplied to data_get_byname")
+            return None
         conn = sqlite3.connect(self._db, uri=True)
         conn.row_factory = dict_factory
         cur = conn.cursor()
@@ -87,6 +100,7 @@ class DataDB(object):
         return None
 
     def data_add(self, data) -> dict:
+        """Insert a data row ."""
         conn = sqlite3.connect(self._db, uri=True)
         cur = conn.cursor()
         query = 'INSERT INTO {} (name, value) VALUES (?,?)'.format(self._table)
@@ -97,6 +111,7 @@ class DataDB(object):
         return data
 
     def data_update(self, data) -> dict:
+        """Update a data row, use id provided in data set."""
         conn = sqlite3.connect(self._db, uri=True)
         cur = conn.cursor()
         query = 'UPDATE {} SET name=?, value=? WHERE id=?'.format(self._table)
@@ -106,11 +121,14 @@ class DataDB(object):
         conn.close()
         return data
 
-    def data_delete(self, id) -> bool:
+    def data_delete(self, key_id) -> bool:
+        """Delete data row indicated by key_id."""
+        if key_id is None:
+            return False
         conn = sqlite3.connect(self._db, uri=True)
         cur = conn.cursor()
         query = 'DELETE FROM {} WHERE id=?'.format(self._table)
-        cur.execute(query, [id, ])
+        cur.execute(query, [key_id, ])
         conn.commit()
         conn.close()
         return True
